Add files via upload
It is an auction web based application in which there are two roles admin and user. Admin can post the product on the website that are available for auctioning and different users can bid the products by login through their account. It is java ,spring framework and hibernate based maven project.